Feature/t 001 criar mapper dto

+++ src/main/java/com/kava/s/CadastroNinja/missao/service/MissaoService.java
@@ -1,44 +1,58 @@
 package com.kava.s.CadastroNinja.missao.service;
 
+import com.kava.s.CadastroNinja.missao.dto.MissaoDTO;
+import com.kava.s.CadastroNinja.missao.mapper.MissaoMapper;
 import com.kava.s.CadastroNinja.missao.models.MissaoModel;
 import com.kava.s.CadastroNinja.missao.repository.MissaoRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class MissaoService {
 
     private MissaoRepository missaoRepository;
+    private MissaoMapper missaoMapper;
 
-    public MissaoService(MissaoRepository missaoRepository) {
+    public MissaoService(MissaoRepository missaoRepository, MissaoMapper missaoMapper) {
         this.missaoRepository = missaoRepository;
+        this.missaoMapper = missaoMapper;
     }
 
     //POST -- Manda uma requisição para criar as missões
-    public MissaoModel criarMissao(MissaoModel missao){
-        return missaoRepository.save(missao);
+    public MissaoDTO criarMissao(MissaoDTO missaoDTO){
+        MissaoModel missao = missaoMapper.map(missaoDTO);
+        missaoRepository.save(missao);
+        return missaoMapper.map(missao);
     }
 
     //GET -- Manda uma requisição para listar todos as missões
-    public List<MissaoModel> listarMissoes(){
-        return missaoRepository.findAll();
+    public List<MissaoDTO> listarMissoes(){
+        List<MissaoModel> missaoModel = missaoRepository.findAll();
+
+        return missaoModel.stream()
+                .map(missaoMapper::map)
+                .collect(Collectors.toList());
     }
 
     //GET -- Manda uma requisição para listar missao por ID
-    public MissaoModel listarMissaoId(Long id){
+    public MissaoDTO listarMissaoId(Long id){
         Optional<MissaoModel> caixaMissao = missaoRepository.findById(id);
 
-        return caixaMissao.orElse(null);
+        return caixaMissao.map(missaoMapper::map).orElse(null);
     }
 
     //PUT -- Manda uma requisição para Alterar uma Missao
-    public MissaoModel alterarMissao(Long id, MissaoModel missaoAtualizada){
-        if(missaoRepository.existsById(id)){
-            missaoAtualizada.setId(id);
+    public MissaoDTO alterarMissao(Long id, MissaoDTO missaoDTO){
+        Optional<MissaoModel> caixaMissao = missaoRepository.findById(id);
 
-            return missaoRepository.save(missaoAtualizada);
+        if(caixaMissao.isPresent()){
+            MissaoModel missao = missaoMapper.map(missaoDTO);
+            missao.setId(id);
+
+            return missaoMapper.map(missao);
         }
 
         return null;
@@ -46,6 +60,10 @@ public class MissaoService {
 
     //DELETE -- Manda uma requisição para deletar uma Missão
     public void deletarMissao(Long id){
-        missaoRepository.deleteById(id);
+        Optional<MissaoModel> caixaMissao = missaoRepository.findById(id);
+
+        if(caixaMissao.isPresent()){
+            missaoRepository.deleteById(id);
+        }
     }
 }
