+++ src/main/java/com/kava/s/CadastroNinja/ninja/service/NinjaService.java
@@ -1,50 +1,68 @@
 package com.kava.s.CadastroNinja.ninja.service;
 
+import com.kava.s.CadastroNinja.ninja.dto.NinjaDTO;
+import com.kava.s.CadastroNinja.ninja.mapper.NinjaMapper;
 import com.kava.s.CadastroNinja.ninja.models.NinjaModel;
 import com.kava.s.CadastroNinja.ninja.repository.NinjaRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class NinjaService {
 
     private NinjaRepository ninjaRepository;
+    private NinjaMapper ninjaMapper;
 
-    public NinjaService(NinjaRepository ninjaRepository) {
+    public NinjaService(NinjaRepository ninjaRepository, NinjaMapper ninjaMapper) {
         this.ninjaRepository = ninjaRepository;
+        this.ninjaMapper = ninjaMapper;
     }
 
     //POST -- Manda uma requisição para criar ninja
-    public NinjaModel criarNinja(NinjaModel ninja){
-        return ninjaRepository.save(ninja);
+    public NinjaDTO criarNinja(NinjaDTO ninjaDTO){
+        NinjaModel ninja = ninjaMapper.map(ninjaDTO);
+        ninjaRepository.save(ninja);
+        return ninjaMapper.map(ninja);
     }
 
     //GET -- Manda uma requisição para listar todos os ninjas
-    public List<NinjaModel> listarNinjas(){
-        return ninjaRepository.findAll();
+    public List<NinjaDTO> listarNinjas(){
+        List<NinjaModel> ninjaModel = ninjaRepository.findAll();
+
+        return ninjaModel.stream()
+                .map(ninjaMapper::map)
+                .collect(Collectors.toList());
     }
 
     //GET -- Manda uma requisição para listar ninja por ID
-    public NinjaModel listarNinjaId(Long id){
+    public NinjaDTO listarNinjaId(Long id){
         Optional<NinjaModel> caixaNinja = ninjaRepository.findById(id);
 
-        return caixaNinja.orElse(null);
+        return caixaNinja.map(ninjaMapper::map).orElse(null);
     }
 
     //PUT -- Manda uma requisição para alterar um ninja
-    public NinjaModel alterarNinja(Long id, NinjaModel ninjaAtualizado){
-        if(ninjaRepository.existsById(id)){
-            ninjaAtualizado.setId(id);
-            return ninjaRepository.save(ninjaAtualizado);
-        }
+    public NinjaDTO alterarNinja(Long id, NinjaDTO ninjaDTO){
+        Optional<NinjaModel> caixaNinja = ninjaRepository.findById(id);
 
+        if(caixaNinja.isPresent()){
+            NinjaModel ninja = ninjaMapper.map(ninjaDTO);
+            ninja.setId(id);
+
+            return ninjaMapper.map(ninja);
+        }
         return null;
     }
 
     //DELETE -- Manda uma requisição para deletar um ninja
     public void deletarNinja(Long id){
-        ninjaRepository.deleteById(id);
+        Optional<NinjaModel> caixaNinja = ninjaRepository.findById(id);
+
+        if(caixaNinja.isPresent()){
+            ninjaRepository.deleteById(id);
+        }
     }
 }
